evidently_monitoring/requirements.txt
evidently
IPython
google.cloud